chore: change folder structure and user service created
